requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
django-diagram==0.0.1
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
graphviz==0.21
pillow==11.3.0
pygraphviz==1.14
PyJWT==2.10.1
sqlparse==0.5.3
